% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obl_potege.R
\name{obl_potege}
\alias{obl_potege}
\title{Funkcja oblicza potege o wykladniku calkowitym}
\usage{
obl_potege(x, y)
}
\arguments{
\item{x}{double}

\item{y}{double}
}
\value{
double
}
\description{
Funkcja oblicza potege o wykladniku calkowitym
}
\examples{
obl_potege(2, 4)
}
